+++ CHANGELOG.md
@@ -7,10 +7,10 @@
 - Fixed ResetShields not re-setting lostShield to false under certain conditions
 - Fixed RemoveMaxEnergy not working as intended
 - Fixed custom AudioClips not loading correctly on Mac OSX
+- Added RuleBookManager for adding custom rulebook sections (see wiki for more info)
 - Added AllModificationInfos, AllModificationTypes, and modification syncing to SlotModificationManager
 - Added additional functionality to SlotModificationManager - Infos now store name and GUID
 - Added rulebook entry support for slot modifications - use extension method .SetRulebook() when adding your slot modification
-- Added RuleBookManager for adding custom rulebook entries (see wiki)
 - Added more shield-related extensions
 - Added some AbilityInfo-related extensions
 - Added ShieldManager.AllShieldAbilities and ShieldManager.AllShieldInfos for easier tracking of custom shield abilities